update dependencies
Cinnamon has (quite a bit ago actually) changed their packaging. Now we
need cinnamon-session instead of cinnamon-desktop for the session
manager, and cinnamon-settings-daemon so our session has all the
required components (and without the settings-daemon, there is little
meaning in using this kind of session at all).

Also, the cinnamon-settings application, which you might need to edit
any kind of cinnamon-related settings (unless you edit bare dconf) now
resides in the cinnamon package, so we recommend that. However, you can
indeed start i3-cinnamon without it, which allows you to get rid of a
bunch of dependencies (and the Nemo file manager in particular).